+++ setup/setup-index.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
 
-# Delete index.php file if it exists
-if [[ -f "index.php" ]]; then
-    rm index.php
-fi
 
 header_content="<?php
 /**
@@ -12,7 +8,7 @@ header_content="<?php
  * @package $php_namespace
  */
 
-defined('WPINC') || die;
+defined( 'ABSPATH' ) || die;
 
 /**
  * Plugin Name: $plugin_name"
@@ -45,35 +41,5 @@ header_content+="\n * Text Domain: $plugin_textdomain
 
 printf "%b" "$header_content" >index.php
 
-# NOTE: Activation hooks need to be inside index.php file or it might not work properly
-# It can fail without error, WordPress is silently failing in case of error
 
-# The code that runs during plugin activation.
-# - includes/Activator.php
-activation_hook_content="\n
-register_activation_hook( __FILE__, function() {
-
-    // On activate do this.
-    \\$php_namespace\Core\Activator::activate();
-});"
-
-printf "%b" "$activation_hook_content" >>index.php
-
-# The code that runs during plugin deactivation.
-# - includes/Deactivator.php
-deactivation_hook_content="\n
-register_deactivation_hook( __FILE__, function () {
-    
-    // On deactivate do that.
-    \\$php_namespace\Core\Deactivator::deactivate();
-});"
-
-printf "%b" "$deactivation_hook_content" >>index.php
-
-# Run plugin, run
-run_content="\n
-include \"run.php\";"
-
-printf "%b" "$run_content" >>index.php
-
-echo "index.php setup completed."
+echo "Plugin header setup completed."